requirements.txt
adal==1.2.2
antlr4-python3-runtime==4.8
applicationinsights==0.11.9
argcomplete==1.11.1
astroid==2.3.3
azure-batch==8.0.0
azure-cli==2.2.0
azure-cli-command-modules-nspkg==2.0.3
azure-cli-core==2.2.0
azure-cli-nspkg==3.0.4
azure-cli-telemetry==1.0.4
azure-common==1.1.25
azure-core==1.3.0
azure-cosmos==3.1.2
azure-datalake-store==0.0.48
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-identity==1.3.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-apimanagement==0.1.0
azure-mgmt-appconfiguration==0.4.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.52.0
azure-mgmt-batch==7.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==4.1.0rc1
azure-mgmt-cognitiveservices==5.0.0
azure-mgmt-compute==11.1.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc10
azure-mgmt-containerservice==8.1.0
azure-mgmt-cosmosdb==0.12.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==2.2.0
azure-mgmt-eventhub==3.0.0
azure-mgmt-hdinsight==1.3.0
azure-mgmt-imagebuilder==0.2.1
azure-mgmt-iotcentral==2.0.0
azure-mgmt-iothub==0.8.2
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==2.1.1
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==1.1.1
azure-mgmt-monitor==0.7.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.7.0
azure-mgmt-network==9.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.4.0
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==2.0.0
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.6.0
azure-mgmt-redis==7.0.0rc1
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==8.0.1
azure-mgmt-search==2.1.0
azure-mgmt-security==0.1.0
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.4.0
azure-mgmt-signalr==0.3.0
azure-mgmt-sql==0.17.0
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==8.0.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==0.44.0
azure-multiapi-storage==0.2.4
azure-nspkg==3.0.2
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azureml-automl-core==1.1.5.1
azureml-core==1.1.5.3
azureml-dataprep==1.3.5
azureml-dataprep-native==14.1.0
azureml-pipeline==1.1.5
azureml-pipeline-core==1.1.5
azureml-pipeline-steps==1.1.5
azureml-sdk==1.1.5.1
azureml-telemetry==1.1.5.3
azureml-train==1.1.5
azureml-train-automl-client==1.1.5.1
azureml-train-core==1.1.5
azureml-train-restclients-hyperdrive==1.1.5
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.8
distro==1.4.0
docker==4.2.0
dotnetcore2==2.1.13
entrypoints==0.3
fabric==2.5.0
flake8==3.7.9
fusepy==3.0.1
humanfriendly==4.18
idna==2.9
invoke==1.4.1
isodate==0.6.0
isort==4.3.21
javaproperties==0.5.1
jeepney==0.4.3
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsmin==2.2.2
jsondiff==1.2.0
jsonpickle==1.3
knack==0.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==2.0.0
msal==1.1.0
msal-extensions==0.1.3
msrest==0.6.11
msrestazure==0.6.2
ndg-httpsclient==0.5.1
numpy==1.18.1
oauthlib==3.1.0
paramiko==2.7.1
pathspec==0.7.0
pbr==5.4.4
portalocker==1.5.2
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pypiwin32==223
pyreadline==2.1
python-dateutil==2.8.1
pytz==2019.1
pywin32==227
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.15.89
scikit-learn==0.22.2.post1
scipy==1.4.1
scp==0.13.2
SecretStorage==3.1.2
six==1.14.0
sklearn==0.0
sshtunnel==0.1.5
tabulate==0.8.6
urllib3==1.25.8
vsts==0.1.25
vsts-cd-manager==1.0.2
websocket-client==0.56.0
wincertstore==0.2
wrapt==1.11.2
xmltodict==0.12.0
